+++ src/index.ts
@@ -1,5 +1,6 @@
 export * from "./command";
 export * from "./command.directive";
 export * from "./command-ref.directive";
+export * from "./command.util";
 export * from "./config";
 export * from "./module";