+++ src/module.ts
@@ -4,12 +4,8 @@ import { CommandDirective } from "./command.directive";
 import { CommandRefDirective } from "./command-ref.directive";
 import { CommandOptions, COMMAND_DEFAULT_CONFIG, COMMAND_CONFIG } from "./config";
 
-/**
- * @internal
- */
-export const _MODULE_CONFIG = new InjectionToken<CommandOptions | (() => CommandOptions)>(
-	"_command-config"
-);
+/** @internal */
+export const _MODULE_CONFIG = new InjectionToken<CommandOptions>("_command-config");
 
 @NgModule({
 	declarations: [CommandDirective, CommandRefDirective],
@@ -23,7 +19,7 @@ export class CommandModule {
 			providers: [
 				{
 					provide: COMMAND_CONFIG,
-					useFactory: _moduleConfigFactory,
+					useFactory: moduleConfigFactory,
 					deps: [_MODULE_CONFIG],
 				},
 				{ provide: _MODULE_CONFIG, useValue: config },
@@ -32,9 +28,7 @@ export class CommandModule {
 	}
 }
 
-/**
- * @internal
- */
-export function _moduleConfigFactory(config: CommandOptions | (() => CommandOptions)) {
+/** @internal */
+export function moduleConfigFactory(config: CommandOptions | (() => CommandOptions)): CommandOptions {
 	return typeof config === "function" ? config() : config;
 }